+++ src/index.ts
@@ -73,6 +73,8 @@ export {
 export { default as sdfuOperations } from './Device/sdfuOperations';
 export { defaultInitPacket, HashType, FwType } from './Device/initPacket';
 
+export { default as describeError } from './logging/describeError';
+
 export const testUtils = {
     dispatchTo,
     render: testRendererForApps,